Messive refactor and rename
- All gears towards build a package
