+++ microfisher_package/src/microfisher/__main__.py
@@ -1,0 +1,5 @@
+import microfisher.microfisher
+
+
+if __name__ == "__main__":
+    microfisher.microfisher.main()
