+++ microfisher_package/tests/test_kreport.py
@@ -1,8 +1,7 @@
 import pytest
 import shlex
 from tests.test_config import setup_args, config
-# from config import Config
-from run_kreport import CentrifugeKReport
+from microfisher.run_kreport import CentrifugeKReport
 
 
 @pytest.mark.usefixtures("setup_args", "config")
